Create utils & services directory, and move form submission logic into those directories
